backend/requirements.txt
fastapi
uvicorn
pdfplumber
google-generativeai
sentence-transformers
torch
networkx
python-dotenv
python-multipart
